+++ src/app/chat/features/ChatInput/ActionBar/config.ts
@@ -5,6 +5,7 @@ import History from './History';
 import ModelSwitch from './ModelSwitch';
 import Temperature from './Temperature';
 import Token from './Token';
+import Tools from './Tools';
 
 export const actionMap = {
   clear: Clear,
@@ -14,6 +15,7 @@ export const actionMap = {
   stt: STT,
   temperature: Temperature,
   token: Token,
+  tools: Tools,
 } as const;
 
 type ActionMap = typeof actionMap;
@@ -24,7 +26,7 @@ type getActionList = (mobile?: boolean) => ActionKeys[];
 
 // we can make these action lists configurable in the future
 export const getLeftActionList: getActionList = (mobile) =>
-  ['model', 'fileUpload', 'temperature', 'history', !mobile && 'stt', 'token'].filter(
+  ['model', 'fileUpload', 'temperature', 'history', !mobile && 'stt', 'tools', 'token'].filter(
     Boolean,
   ) as ActionKeys[];